rename pomelo to kyubeydex
